UserController

+++ src/main/java/group5/hotelms/controller/UserControllerImpl.java
@@ -1,11 +1,15 @@
 package group5.hotelms.controller;
 
+import group5.hotelms.dao.HotelDAO;
+import group5.hotelms.dao.HotelDAOImpl;
 import group5.hotelms.dao.UserDAO;
 import group5.hotelms.dao.UserDAOImpl;
 import group5.hotelms.model.Data;
+import group5.hotelms.model.Hotel;
 import group5.hotelms.model.User;
-import group5.hotelms.util.DataLoader;
-import group5.hotelms.util.DataLoaderImpl;
+
+import java.util.HashSet;
+import java.util.Set;
 
 
 /**
@@ -14,8 +18,10 @@ import group5.hotelms.util.DataLoaderImpl;
 public class UserControllerImpl implements UserController {
 
 
-    private DataLoader dataLoader = new DataLoaderImpl();
-    private UserDAO userDAO = new UserDAOImpl(new Data(dataLoader.getHotels(), dataLoader.getUsers()));
+    private UserDAO userDAO = new UserDAOImpl();
+    private HotelDAO hotelDAO = new HotelDAOImpl();
+    private Set<Hotel> hotels = hotelDAO.getAllHotels();
+    private Set<User> users = new HashSet<>(Data.getUsers());
 
 
     public boolean register(User user) {
@@ -24,11 +30,26 @@ public class UserControllerImpl implements UserController {
         return userDAO.saveUser(user);
     }
 
+    /**
+     * Delete user, check the room reservation by the user. Check the presence of the user in the database
+     *
+     * @param user
+     * @return
+     */
+
     public boolean remove(User user) {
 
+        Set<Hotel> hotelsWithUser = (Set<Hotel>) hotels.stream().filter(h -> h.getRooms().stream().anyMatch(r -> r.getUser().getLogin().equals(user.getLogin())));
 
+        if (!hotelsWithUser.isEmpty() || !users.stream().equals(user)) {
+
+
+            return false;
+        }
         return userDAO.deleteUser(user);
+
     }
+
 
     public boolean edit(User user) {
